Organiza testes, cobre todos endpoints e atualiza documentação
