+++ swagger.json
@@ -234,6 +234,38 @@
         }
       }
     },
+    "/user": {
+      "patch": {
+        "summary": "Atualiza dados do usuário autenticado",
+        "security": [ { "bearerAuth": [] } ],
+        "requestBody": {
+          "required": true,
+          "content": {
+            "application/json": {
+              "schema": {
+                "type": "object",
+                "properties": {
+                  "password": {
+                    "type": "string",
+                    "minLength": 12,
+                    "maxLength": 16,
+                    "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*()_+\\-=[\\]{};':\\\"|,.<>/?]).{12,16}$"
+                  }
+                },
+                "required": ["password"]
+              }
+            }
+          }
+        },
+        "responses": {
+          "200": { "description": "Usuário atualizado com sucesso." },
+          "400": { "description": "Dados inválidos." },
+          "401": { "description": "Token não fornecido." },
+          "403": { "description": "Token inválido." },
+          "404": { "description": "Usuário não encontrado." }
+        }
+      }
+    },
     "/admin/user": {
       "patch": {
         "summary": "Admin: Atualiza nome ou senha de qualquer usuário",
